[CK TILE] Fix gemm_bquant gaps: preshuffle_b support, layout restrictions, and validation improvements
- Fix static_cast<float>() in compare_bquant for correct max element computation
- Enable preshuffle_b=true in all config files (default, CI, user_provided)
- Add layout restriction: skip b_preshuffle_quant=True for RowMajor B layouts (rrr, crr)
  since PreshuffleQuant only supports ColumnMajor BQ layout
- Add pipeline constants (GEMM_BQUANT/ROWCOL/ABQUANT/TENSOR_PIPELINES) to gemm_validation_utils.py
- Register gemm_abquant, gemm_tensor, gemm_rowcol subdirectories in block_scale_gemm CMakeLists.txt

+++ projects/composablekernel/tile_engine/ops/gemm/block_scale_gemm/gemm_bquant/gemm_bquant_benchmark.hpp
@@ -181,12 +181,12 @@ bool compare_bquant(std::string instanceName,
                     ck_tile::HostTensor<CDataType>& c_m_n_host_result)
 {
     const float max_accumulated_value =
-        std::abs(*std::max_element(c_m_n_host_result.mData.begin(),
-                                   c_m_n_host_result.mData.end(),
-                                   [](const auto& a, const auto& b) {
-                                       return std::abs(static_cast<float>(a)) <
-                                              std::abs(static_cast<float>(b));
-                                   }));
+        std::abs(static_cast<float>(*std::max_element(c_m_n_host_result.mData.begin(),
+                                                      c_m_n_host_result.mData.end(),
+                                                      [](const auto& a, const auto& b) {
+                                                          return std::abs(static_cast<float>(a)) <
+                                                                 std::abs(static_cast<float>(b));
+                                                      })));
     const auto rtol_atol =
         calculate_rtol_atol_bquant<ADataType, BQDataType, BDataType, AccDataType, CDataType>(
             K, kbatch, max_accumulated_value);
